QT: la ventana es escalable
fix sdltesting con las nuevas funciones.

+++ src/client/sdl_ball.cpp
@@ -9,7 +9,7 @@ void sdl_ball::render(SDL2pp::Renderer &renderer) {
     renderer.Copy(texture_ball, SDL2pp::NullOpt, SDL2pp::Rect(x, y, radius*2, radius*2), angle, SDL2pp::NullOpt, SDL_FLIP_NONE);
 }
 
-void sdl_ball::update(int _x, int _y, int _angle, int _radius) {
+void sdl_ball::update(int _x, int _y, double _angle, int _radius) {
     this->x = _x;
     this->y = _y;
     this->angle = _angle;
