+++ src/client/sdl_player.h
@@ -15,12 +15,12 @@ private:
     bool jumping;
     int x;
     int y;
-    int angle;
+    double angle;
     int size_w;
     int size_h;
 public:
     explicit sdl_player(SDL2pp::Renderer &renderer);
-    void update(int x, int y, int angle, int dt, bool isMoving, bool isJumping, bool onTurbo);
+    void update(int x, int y, double angle, float dt, bool isMoving, bool isJumping, bool onTurbo);
     void render(SDL2pp::Renderer &renderer);
     ~sdl_player();
